corrected citation from `dierksen2015` to `@dirksen2015`
There are two missing citations (rendered as **???**) in bookdown. The first is in section 16.2.5--_however_ the citation appears correct, so unsure why it is not rendering. In the second, there appears to be a typo. I corrected it (I think 😉) to match the key in the book.bib file. I also flagged this with a new issue.