Update 03-data-scientist-in-edu.Rmd
Underlying structure: is it a defined phrase, a norm of stats, etc.? If not, consider rewording or defining the phrase in context of the book. 
Explor- Root: Used 11 times in chapter 3. 7 times used in a span of 5 chapters. Edited to be more specific in word usage.

Overarching Comment: Frequent use of qualifying language (we see this as, we think, we believe, etc.). Cheapens the value of your statements and is unnecessary.