Update 16-teaching-data-science.Rmd
Line 248: This paragraph is hard to comprehend. Your first sentence (rewritten in the edits) talks about variables being subjective or non-representative of the true population, but the second half of the paragraph talks about interpretation of data. I'm not 100% sold on the edits I made so consider what you are trying to say and rewrite for clarity as needed.